Support keeping segments together
Without any precautions, it could happen that a line would break
exactly in between two segments while this isn't always allowed. To
fix this, in css-to-obfl we insert a ZWNBSP character in between
segments that should be kept together.

Upon encountering such a ZWNBSP character,
AbstractBrailleTranslator.nextTranslatedRow does not include the last
word of the segment. It does this because it assumes that the first
word of the next segment is too wide to fit on the same line. It has
to make this assumption because it does not know anything about the
surrounding segments.

By making use of the other method, getTranslatedRemainder, Dotify can
find the first segment that does not fit completely on the line, and
look for break points starting from that segments and going back
towards the first.

For more info see the discussion in issues

- https://github.com/sbsdev/pipeline-mod-sbs/issues/63,
- https://github.com/brailleapps/dotify.api/issues/13 and
- https://github.com/brailleapps/dotify.formatter.impl/pull/45